Add nameImprovement parameter
