+++ functions/.eslintrc.js
@@ -26,5 +26,6 @@ module.exports = {
     'object-curly-spacing': ['warn', 'always'],
     'require-jsdoc': 'off',
     'semi': ['error', 'never'],
+    'no-unused-vars': ['warn', { 'argsIgnorePattern': '_' }],
   },
 }
