+++ functions/src/index.ts
@@ -1,85 +1,43 @@
-import * as vision from '@google-cloud/vision'
 import 'firebase-functions'
 import * as functions from 'firebase-functions'
 import { firestoreBackup } from './admin'
-import { Product } from './types/products'
-import { defaultStore, walmart } from './types/stores'
-import { verticalSegment } from './utils'
+import { analyzeReceipt } from './functions/analyzeReceipt'
 
 export const scheduledBackup = firestoreBackup
 
 export const setTimestampOnPaymentCreation = functions.firestore
     .document('payments/{paymentId}')
-    .onCreate(async (snap, context) => {
+    .onCreate(async (snap, _) => {
       await snap.ref.update({ timeCreated: snap.createTime })
     })
 
 export const getReceiptDataTest = functions
     .runWith({ timeoutSeconds: 300 })
-    .https.onRequest(
-        async (request, response) => {
-          const receiptUrl = request.query.receiptUrl
-          const isWalmart = request.query.isWalmart
-          if (!receiptUrl) {
-            response.status(400).send('Parameter receiptUrl is required')
-          }
+    .https.onRequest(async (request, response) => {
+      const { receiptUrl, isWalmart, withNameImprovement } = request.query
 
-          const result = await analyzeReceipt(
-            receiptUrl as string,
-            isWalmart === 'true'
-          )
-          response.send(result)
-        }
-    )
+      if (!receiptUrl) {
+        response.status(400).send('Parameter receiptUrl is required')
+      }
+
+      const result = await analyzeReceipt(
+        receiptUrl as string,
+        isWalmart === 'true',
+        withNameImprovement === 'true'
+      )
+      response.send(result)
+    })
 
 export const getReceiptData = functions
     .runWith({ timeoutSeconds: 300 })
-    .https.onCall(async (data, context) => {
+    .https.onCall(async (data, _) => {
       if (!data.receiptUrl) {
         throw Error('The parameter receiptUrl is required.')
       }
 
-      return analyzeReceipt(data.receiptUrl, data.isWalmart)
+      return analyzeReceipt(
+          data.receiptUrl,
+          data.isWalmart,
+          data.withNameImprovement
+      )
     })
-
-async function analyzeReceipt(
-    receiptUrl: string,
-    isWalmart: boolean
-): Promise<Product[]> {
-  const client = new vision.ImageAnnotatorClient()
-
-  const [result] = await client.textDetection(receiptUrl)
-
-  // first element contains information about all lines
-  const labels = result.textAnnotations?.slice(1) ?? []
-
-  type LabelBox = { p1: number; p2: number; description: string }
-  const lines: LabelBox[][] = []
-
-  labels.forEach((label) => {
-    const labelSegment = verticalSegment(label)
-    const center = (labelSegment.p1 + labelSegment.p2) / 2
-    const labelBox = { ...labelSegment, description: label.description ?? '' }
-
-    for (const line of lines) {
-      if (line[0].p1 < center && center < line[0].p2) {
-        line.push(labelBox)
-        return
-      }
-    }
-
-    lines.push([labelBox])
-  })
-
-  const rows = lines.map((line) => line.map((label) => label.description))
-  console.log({ rows })
-
-  const store = isWalmart ? walmart : defaultStore
-
-  let products = store.normalize(rows)
-  products = store.filter(products)
-  products = store.merge(products)
-  products = await store.improveNaming(products)
-
-  return products
-}
